requirements.txt
# Data processing and analysis
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0

# Database
SQLAlchemy>=2.0.21
mysql-connector-python>=8.1.0
pymysql>=1.1.0

# Web API for dynamic dashboard
Flask>=3.0.0
Flask-CORS>=4.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.16.1
reportlab>=4.0.7

# Jupyter
jupyter>=1.0.0
notebook>=7.0.3

# Testing
pytest>=7.4.2

# Configuration
PyYAML>=6.0.1

# Type hints
mypy>=1.5.1
types-PyYAML>=6.0.12.11
